fix(php-sdk): Daemon/GrpcClient — distinguish recv timeout vs stream close
Two interacting bugs surfaced once Swoole was installed locally and the
Swoole/* tests actually ran:

1. Swoole\Coroutine\Channel::pop(0.0) blocks indefinitely (Swoole
   treats timeout=0 as "no timeout", not "non-blocking poll"). The
   Daemon's steady-state outbound drain used 0.0 → first iteration
   hung forever. Fix: gate the drain on outbound->length() > 0 before
   calling pop, and use a tiny positive timeout (0.001s) on the pop
   itself. Same fix on the shutdown drain() loop.

2. GrpcClient::recv() returned null on both timeout AND stream close.
   The Daemon's steady-state loop continues on null (it's just a
   timeout) and would loop forever once the test fake's scripted
   inbound list ran out. Fix: recv() now throws ConnectorException
   on stream close (detected via http2->errCode != 0 or false-return
   well within the deadline window). Daemon's steady-state loop
   catches the exception and breaks cleanly.

3. Test fake GrpcClient in DaemonTest now mirrors the real contract:
   throws ConnectorException on a scripted null entry (EOF marker)
   or once the script is exhausted. Daemon test passes in <0.5s
   instead of hanging.

4. GrpcClientTest's anonymous fakes had read(int $timeout) — real
   Swoole signature is read(float $timeout). Fixed all 3 fakes.

5. Daemon's drainGraceSeconds default stays at 30s for production,
   but DaemonTest now constructs the daemon with drainGraceSeconds=1
   so the test exits in ~1s even in the empty-channel path.

Full unit suite: 62 passed, 1 skipped (intentional integration
placeholder), 0 failed. PHPStan clean.

+++ tests/Unit/Swoole/GrpcClientTest.php
@@ -23,7 +23,7 @@ it('open() initializes the stream with the right headers', function () {
             $this->streamId = 7;
             return 7;
         }
-        public function read(int $timeout = -1) { return false; }
+        public function read(float $timeout = -1) { return false; }
         public function write(int $streamId, string $data, bool $end = false): bool { return true; }
         public function close(): bool { return true; }
     };
@@ -49,7 +49,7 @@ it('send() frames a ConnectorMsg and writes it as DATA', function () {
         public bool $connected = false;
         public function connect(): bool { return true; }
         public function send($req): int|false { return 1; }
-        public function read(int $timeout = -1) { return false; }
+        public function read(float $timeout = -1) { return false; }
         public function write(int $streamId, string $data, bool $end = false): bool {
             $this->captured[] = $data; return true;
         }
@@ -85,7 +85,7 @@ it('recv() reads a DATA frame and decodes it as HubMsg', function () {
         public function connect(): bool { return true; }
         public function send($req): int|false { return 1; }
         public function write(int $sid, string $d, bool $e = false): bool { return true; }
-        public function read(int $timeout = -1) {
+        public function read(float $timeout = -1) {
             if ($this->reads++ > 0) return false;
             $r = new \stdClass();
             $r->data = $this->framed;
